pyproject.toml
[project]
name = "recipe-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=0.6.0",
    "langchain>=1.2.10",
    "langchain-core>=1.2.14",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.9",
    "pypdf>=5.0.0",
    "langfuse>=2.0.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]
